+++ builtin/aws/alb/releaser.go
@@ -11,10 +11,10 @@ import (
 	"github.com/aws/aws-sdk-go/service/elbv2"
 	"github.com/aws/aws-sdk-go/service/route53"
 	"github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/waypoint/builtin/aws/utils"
 	"github.com/hashicorp/waypoint-plugin-sdk/component"
 	"github.com/hashicorp/waypoint-plugin-sdk/docs"
 	"github.com/hashicorp/waypoint-plugin-sdk/terminal"
+	"github.com/hashicorp/waypoint/builtin/aws/utils"
 )
 
 type Releaser struct {
@@ -39,7 +39,10 @@ func (r *Releaser) Release(
 	ui terminal.UI,
 	target *TargetGroup,
 ) (*Release, error) {
-	sess := session.New(aws.NewConfig().WithRegion(target.Region))
+	sess, err := session.NewSession(aws.NewConfig().WithRegion(target.Region))
+	if err != nil {
+		return nil, err
+	}
 	elbsrv := elbv2.New(sess)
 
 	lbName := r.config.Name