+++ builtin/aws/ecr/registry.go
@@ -21,7 +21,7 @@ import (
 	"github.com/mattn/go-isatty"
 )
 
-// Registry represents access to a Docker registry.
+// Registry represents access to an AWS registry.
 type Registry struct {
 	config Config
 }
@@ -54,8 +54,10 @@ func (r *Registry) Push(
 
 	cli.NegotiateAPIVersion(ctx)
 
-	sess := session.New(aws.NewConfig().WithRegion(r.config.Region))
-
+	sess, err := session.NewSession(aws.NewConfig().WithRegion(r.config.Region))
+	if err != nil {
+		return nil, err
+	}
 	svc := ecr.New(sess)
 
 	repOut, err := svc.DescribeRepositories(&ecr.DescribeRepositoriesInput{